+++ src/confluence/pif_maker.py
@@ -1,4 +1,14 @@
+def create_pif(headers, row)
 
+    chemical_system = ChemicalSystem()
+    chemical_system.chemical_formula = 'MgO2'
+
+    band_gap = Property()
+    band_gap.name = 'Band gap'
+    band_gap.scalars = 7.8
+    band_gap.units = 'eV'
+
+    chemical_system.properties = band_gap
 
 
 def convert(files=[], **kwargs):
